requirements.txt
discord.py>=2.3,<3.0
aiohttp>=3.9,<4.0
aiosqlite>=0.19,<1.0
python-dotenv>=1.0,<2.0
